og hopon/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.5
python-socketio==5.10.0
python-engineio==4.8.0
PyJWT==2.8.0
authlib==1.3.0
requests==2.31.0
cryptography==41.0.7
python-dotenv==1.0.0
marshmallow==3.20.1
gevent==23.9.1
gevent-websocket==0.10.1
redis==5.0.1
Werkzeug==3.0.1
